Move betty.plugins.jinja_filter to betty.jinja_filters
