+++ betty/jinja_filters/format_degrees.py
