+++ betty/jinja_filters/image_resize_cover.py
@@ -22,9 +22,9 @@ from betty.hashid import hashid_file_meta
 from betty.image import image_file_path_format, resize_cover
 from betty.jinja import context_document
 from betty.jinja.filter import JinjaFilter, JinjaFilterDefinition
+from betty.jinja_filters.file import File as FileFilter
 from betty.media_types.svg import SVG
 from betty.os import _link_or_copy
-from betty.plugins.jinja_filter.file import File as FileFilter
 from betty.project import Project
 
 if TYPE_CHECKING:
