+++ betty/jinja_filters/json_dump.py
