+++ betty/jinja_filters/negotiate_has_dates.py
@@ -10,7 +10,7 @@ from typing import TYPE_CHECKING, final
 from jinja2 import pass_context
 
 from betty.jinja.filter import JinjaFilter, JinjaFilterDefinition
-from betty.plugins.jinja_filter.select_has_dates import SelectHasDates
+from betty.jinja_filters.select_has_dates import SelectHasDates
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
