+++ betty/plugins/content/tree.py
@@ -10,9 +10,9 @@ from betty.document import Document
 from betty.entities.person import Person
 from betty.extensions.webpack import Webpack
 from betty.factory import Manufacturable
+from betty.jinja_filters.webpack_entry_point_js import WebpackEntryPointJs
 from betty.locale.localizable.gettext import _
 from betty.plugins.content.template import Template, TemplateBuild
-from betty.plugins.jinja_filter.webpack_entry_point_js import WebpackEntryPointJs
 from betty.project import Project
