+++ betty/tests/jinja_filters/test_format_degrees.py
