+++ betty/tests/jinja_filters/test_image_resize_cover.py
