+++ pyproject.toml
@@ -231,21 +231,21 @@ betty = 'betty.console:main_standalone'
 'woman' = 'betty.genders.woman:Woman'
 
 [project.entry-points.'betty.jinja_filter']
-'build-content' = 'betty.plugins.jinja_filter.build_content:BuildContent'
-'file' = 'betty.plugins.jinja_filter.file:File'
-'format-datetime-datetime' = 'betty.plugins.jinja_filter.format_datetime_datetime:FormatDatetimeDatetime'
-'format-degrees' = 'betty.plugins.jinja_filter.format_degrees:FormatDegrees'
-'html-lang' = 'betty.plugins.jinja_filter.html_lang:HtmlLang'
-'image-resize-cover' = 'betty.plugins.jinja_filter.image_resize_cover:ImageResizeCover'
-'json-dump' = 'betty.plugins.jinja_filter.json_dump:JsonDump'
-'json-load' = 'betty.plugins.jinja_filter.json_load:JsonLoad'
-'localize' = 'betty.plugins.jinja_filter.localize:Localize'
-'negotiate-has-dates' = 'betty.plugins.jinja_filter.negotiate_has_dates:NegotiateHasDates'
-'select-has-dates' = 'betty.plugins.jinja_filter.select_has_dates:SelectHasDates'
-'to-language-tag' = 'betty.plugins.jinja_filter.to_language_tag:ToLanguageTag'
-'unique' = 'betty.plugins.jinja_filter.unique:Unique'
-'url' = 'betty.plugins.jinja_filter.url:Url'
-'webpack-entry-point-js' = 'betty.plugins.jinja_filter.webpack_entry_point_js:WebpackEntryPointJs'
+'build-content' = 'betty.jinja_filters.build_content:BuildContent'
+'file' = 'betty.jinja_filters.file:File'
+'format-datetime-datetime' = 'betty.jinja_filters.format_datetime_datetime:FormatDatetimeDatetime'
+'format-degrees' = 'betty.jinja_filters.format_degrees:FormatDegrees'
+'html-lang' = 'betty.jinja_filters.html_lang:HtmlLang'
+'image-resize-cover' = 'betty.jinja_filters.image_resize_cover:ImageResizeCover'
+'json-dump' = 'betty.jinja_filters.json_dump:JsonDump'
+'json-load' = 'betty.jinja_filters.json_load:JsonLoad'
+'localize' = 'betty.jinja_filters.localize:Localize'
+'negotiate-has-dates' = 'betty.jinja_filters.negotiate_has_dates:NegotiateHasDates'
+'select-has-dates' = 'betty.jinja_filters.select_has_dates:SelectHasDates'
+'to-language-tag' = 'betty.jinja_filters.to_language_tag:ToLanguageTag'
+'unique' = 'betty.jinja_filters.unique:Unique'
+'url' = 'betty.jinja_filters.url:Url'
+'webpack-entry-point-js' = 'betty.jinja_filters.webpack_entry_point_js:WebpackEntryPointJs'
 
 [project.entry-points.'betty.jinja_test']
 'has-file-references' = 'betty.jinja_tests.has_file_references:HasFileReferences'
